Update Rust Network IO

+++ src/main/rust/src/runtime/ObjectReader.rs
@@ -25,7 +25,7 @@ pub struct ObjectReader {
     objectLength: Vec<u32>,
     pagePosition: Vec<u64>,
     currentPagePosition: u64,
-    fileSize: u64,
+    fileSize: i64,
     networkPageCount: u32,
 }
 
@@ -37,7 +37,7 @@ impl ObjectReader {
         let mut tmpPagePosition: Vec<u64> = vec![];
 
         let mut tmpFile = File::open(fname).unwrap();
-        let tmpFileSize = tmpFile.metadata().unwrap().len();
+        let tmpFileSize = tmpFile.metadata().unwrap().len() as i64;
         let tmpNetworkPageCount = ((tmpFileSize as f64 / Const::NETWORK_PAGESIZE as f64) as f64).ceil() as u32;
         tmpFile.flush().unwrap();
 
@@ -215,13 +215,13 @@ impl ObjectReader {
     }
 
     pub fn readAllPages(&mut self, pages:&mut Vec<Vec<u8>>, pagesSize:&mut Vec<i32>){
-        for i in 0..self.networkPageCount as u64{
+        for i in 0..self.networkPageCount as i64{
             let mut pageSize:i32 = Const::NETWORK_PAGESIZE as i32;
-            if self.fileSize - (i+1) * (Const::NETWORK_PAGESIZE as u64) < 0 {
-                pageSize = (self.fileSize - i * Const::NETWORK_PAGESIZE as u64) as i32;
+            if self.fileSize - (i+1) * (Const::NETWORK_PAGESIZE as i64) < 0 {
+                pageSize = (self.fileSize - i * Const::NETWORK_PAGESIZE as i64) as i32;
             }
             pagesSize.push(pageSize);
-            let newPosition = i * pageSize as u64;
+            let newPosition = (i * pageSize as i64) as u64;
             self.inStreamRegularFile.seek(SeekFrom::Start(newPosition));
             let mut reader = BufReader::with_capacity(pageSize as usize, &self.inStreamRegularFile);
             let buffer = reader.fill_buf().unwrap();
